Allow minimum lag index to be specified.
Previously the minimum lag index was assumed/forced to be zero.
While theoretically this is a good assumption, it may not be correct
due to data processing like statics corrections. Now a positve or
negative minimum lag index may be specified explicity. The maximum
lag index is still computed from a guess of the average Vp/Vs ratio.